requirements.txt
astroid==2.4.2
boto3==1.15.6
botocore==1.18.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
isort==5.5.3
jmespath==0.10.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
networkx==2.5
numpy==1.19.2
pika==1.1.0
Pillow==7.2.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
PyWavelets==1.1.1
requests==2.24.0
s3transfer==0.3.3
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tifffile==2020.9.22
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
